Fix type annotation errors and add missing dependencies
- Fix optional type hints (str | None) in backtest_runner.py
- Fix field default_factory in lean-ops-mcp config.py
- Add type ignores for Streamlit and pandas compatibility
- Add connection None checks in learning-store database.py
- Convert numpy floats to Python floats in patterns.py and learner.py
- Add proper type annotations for mixed-type lists
- Install adaptive-agent package with apscheduler dependency
- Update tests pyproject.toml with proper package structure
- All 44 unit tests now passing

+++ learning-store/src/learning_store/learner.py
@@ -248,7 +248,7 @@ class AdaptiveLearner:
             # Update statistics
             pattern.sample_size += len(matching_trades)
             pattern.win_rate = (pattern.win_rate + new_win_rate) / 2  # Moving average
-            pattern.avg_pnl = (pattern.avg_pnl + new_avg_pnl) / 2
+            pattern.avg_pnl = float((pattern.avg_pnl + new_avg_pnl) / 2)
             
             update_pattern(pattern)
